Replaced bfloat16 to float16

+++ agents.py
@@ -52,7 +52,7 @@ class ActorCriticAgent(nn.Module):
         self.lambd = lambd
         self.entropy_coef = entropy_coef
         self.use_amp = True
-        self.tensor_dtype = torch.bfloat16 if self.use_amp else torch.float32
+        self.tensor_dtype = torch.float16 if self.use_amp else torch.float32
 
         self.symlog_twohot_loss = SymLogTwoHotLoss(255, -20, 20)
 
@@ -125,7 +125,7 @@ class ActorCriticAgent(nn.Module):
     def sample(self, latent, greedy=False):
         self.eval()
         with torch.autocast(
-            device_type=device.type, dtype=torch.bfloat16, enabled=self.use_amp
+            device_type=device.type, dtype=torch.float16, enabled=self.use_amp
         ):
             logits = self.policy(latent)
             dist = distributions.Categorical(logits=logits)
@@ -147,7 +147,7 @@ class ActorCriticAgent(nn.Module):
         """
         self.train()
         with torch.autocast(
-            device_type=device.type, dtype=torch.bfloat16, enabled=self.use_amp
+            device_type=device.type, dtype=torch.float16, enabled=self.use_amp
         ):
             logits, raw_value = self.get_logits_raw_value(latent)
             dist = distributions.Categorical(logits=logits[:, :-1])
